price/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
click==8.1.7
colorama==0.4.6
fastapi==0.110.0
grpcio==1.62.0
grpcio-tools==1.62.0
h11==0.14.0
idna==3.6
protobuf==4.25.3
pydantic==2.6.3
pydantic_core==2.16.3
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.27.1
